backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.1
numpy==1.26.4
scikit-learn==1.5.1
shap==0.46.0
pdfplumber==0.11.4
pymongo==4.8.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
openpyxl==3.1.5